Xシェア機能を追加
