+++ front/src/components/features/illusts/index.ts
@@ -1,4 +1,5 @@
 import Illust from "./illust";
 import IllustCarousel from "./illustCarousel";
+import XShare from "./xShare";
 
-export { Illust, IllustCarousel };
+export { Illust, IllustCarousel, XShare };
