+++ front/src/components/ui/index.ts
@@ -2,10 +2,7 @@
 import { TransitionsModal, SearchModal, RequiredLoginModal } from "./modal";
 import H2 from "./headTag";
 import ToggleSort from "./toggleSort";
-import {
-  IconButtonList,
-  FixedIconButtonList,
-} from "./iconsButton/iconButtonList";
+import { IconButtonList, FixedIconButtonList } from "./iconsButton";
 import Collapse from "./collapse";
 import Pagination from "./pagination";
 import LoginLink from "./loginLink";
